added logic and pop-up's for User
